Add "In the News" section to expert pages
