+++ people/search.py
@@ -1,52 +1,9 @@
-from core.models import ContentPage
 from wagtail.search.backends import get_search_backend
 from wagtail.search.backends.elasticsearch7 import (
     Elasticsearch7SearchResults
 )
 
 from .models import PersonPage
-
-
-class CIGIOnlineExpertLatestActivitySearchQueryCompiler:
-    def __init__(self, expert_id):
-        self.expert_id = expert_id
-
-    @property
-    def queryset(self):
-        return ContentPage.objects.live()
-
-    def get_query(self):
-        return {
-            "bool": {
-                "filter": [{
-                    "term": {
-                        "live_filter": True,
-                    },
-                }, {
-                    "terms": {
-                        "content_type": ["core.ContentPage"],
-                    },
-                }, {
-                    "terms": {
-                        "core_contentpage__related_people_ids_filter": [self.expert_id],
-                    },
-                }],
-            },
-        }
-
-    def get_sort(self):
-        return [{
-            "core_contentpage__publishing_date_filter": {
-                "order": "desc",
-            },
-        }]
-
-
-def expert_latest_activity_search(expert_id):
-    return Elasticsearch7SearchResults(
-        get_search_backend(),
-        CIGIOnlineExpertLatestActivitySearchQueryCompiler(expert_id),
-    )
 
 
 class CIGIOnlineExpertsSearchQueryCompiler:
